Rewrite $notice to split display and alt text
Aside from fixing a minor spelling mistake, I changed the $notice variable to hold an array rather than a really long string. The dp_alt_products function then wraps the first array item (the display text) in an h4, and checks the array's length to see if a second array item (the alt product text) exists. If so, it wraps that in another h4 with a different class and includes it in the $dp_display_notice variable, which is echoed to the page.